v0.1.2 - Beta
This is a beta version, needs polishing, testing and is not complete, you SHOULD NOT use this as a production app.

This update starts the addition of a local database so the app works even offline, synchronizing to the external database after ethernet connection is restored.
This update also adds some new UI features to make it more user friendly, mainly when the user punches his card.

Also fixed some other things, such as the Toaster not working properly and etc.

Next update I'll add a proper README and MAYBE some translations to EN-US.